refactor: extract useThreadActions composable from MailboxView
Move the list/reading-pane thread action layer (7 mail-level resources, the
mail-id helpers, optimistic/undo/toast logic, toolbar option computeds, the
junk/delete dialog, and all handle* functions) out of MailboxView.vue into
src/utils/useThreadActions.ts. MailboxView drops ~510 lines and calls the
composable, passing reactive mailbox/threadID and the few view-owned callbacks.
Behavior is unchanged (undoAction is a shared singleton, so undo still works
across the boundary).

(cherry picked from commit 6e6f47dcb579593300c614124ad2d70cc9fd15e2)
